# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Social Media API
  description: |-
    An open-source RESTful API by [NixLab Technologies](https://nixlab.co.in) developed using NodeJS, ExpressJS and MongoDB to integrate backend and frontend with ease.

    You can now help us improve the API whether it's by making changes to the definition itself or to the code.
    That way, with time, we can improve the API in general, and expose some of the new features in OAS3.

    Some useful links:
    - [Repository (GitHub)](https://github.com/nixrajput/social-media-api-nodejs)
    - [API definition](https://github.com/nixrajput/social-media-api-nodejs/blob/master/openapi.yaml)

  termsOfService: https://nixlab.co.in/terms
  contact:
    email: nixlab.in@gmail.com
  license:
    name: GPL 3.0
    url: https://github.com/nixrajput/social-media-api-nodejs/blob/master/LICENSE.md
  version: 1.0.0
...
servers:
  - url: https://social-api.nixlab.co.in/api/v1
...
paths:
...
  /forgot-password:
    post:
      tags:
        - auth
      summary: Send OTP for password reset
      description: Send OTP for password reset
      operationId: forgotPassword
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: "test@email.com"
        required: true
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                    example: "otp has been sent successfully"
        "400":
          description: error occurred
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                    example: "email is invalid"
        "404":
          description: user not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                    example: "user not found"
  /reset-password:
    post:
      tags:
        - auth
      summary: Reset your password using OTP
      description: Reset your password using OTP
      operationId: resetPassword
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                otp:
                  type: string
                  example: "0123456"
                newPassword:
                  type: string
                  example: "password"
                confirmPassword:
                  type: string
                  example: "password"
        required: true
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                    example: "password has been reset successfully"
        "400":
          description: error occurred
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                    example: "otp is invalid"
        "404":
          description: user not found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                    example: "user not found"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...